Fatwa Video

+++ app/src/main/java/com/ch/al_mirats/utils/CurrencyFormat.kt
@@ -4,7 +4,7 @@ import android.icu.text.NumberFormat
 import android.icu.util.Currency
 import kotlin.math.roundToInt
 
-fun Int.toCurrencyFormat():String{
+fun Long.toCurrencyFormat():String{
     val format: NumberFormat = NumberFormat.getCurrencyInstance()
     format.maximumFractionDigits = 0
     format.currency = Currency.getInstance("IDR")
